refactor: 상수 분류 안하셔서 기능별로 리팩토링
